just a fill changes while testing things

+++ Engine/ivsEngine.h
@@ -53,6 +53,6 @@ public:
 	inline static	glm::vec3	cameraUp = { 0.0f, 1.0f, 0.0f };
 	inline static	glm::vec4	uiColor = { 1.0f, 0.5f, 0.0f, 1.0f };
 	inline static	int			padding = 10;
-	inline static	int			maxVertexperTerrain = 10;
-	inline static	int			terrainSize  = 10;
+	inline static	float		maxVertexperTerrain = 10;
+	inline static	float		terrainSize  = 10;
 };
